stage

+++ src/TD.Komercijalno/TD.Komercijalno.Contracts/Dtos/Magacini/MagacinDto.cs
@@ -4,7 +4,7 @@ namespace TD.Komercijalno.Contracts.Dtos.Magacini
 {
     public class MagacinDto
     {
-        public short MagacinId { get; set; }
+        public int MagacinId { get; set; }
         public string Naziv { get; set; }
         public string MtId { get; set; }
         public short VodiSe { get; set; }
